refactor: type files move in to src/workbench/extensions/manager directory
