+++ src/workbench/extensions/manager/components/manager/PackStatusMessage.vue
@@ -20,7 +20,7 @@ import Message from 'primevue/message'
 import { computed, inject } from 'vue'
 
 import type { components } from '@/types/comfyRegistryTypes'
-import { ImportFailedKey } from '@/types/importFailedTypes'
+import { ImportFailedKey } from '@/workbench/extensions/manager/types/importFailedTypes'
 
 type PackVersionStatus = components['schemas']['NodeVersionStatus']
 type PackStatus = components['schemas']['NodeStatus']
