+++ tests-ui/tests/stores/conflictDetectionStore.test.ts
@@ -1,8 +1,8 @@
 import { createPinia, setActivePinia } from 'pinia'
 import { beforeEach, describe, expect, it } from 'vitest'
 
-import type { ConflictDetectionResult } from '@/types/conflictDetectionTypes'
 import { useConflictDetectionStore } from '@/workbench/extensions/manager/stores/conflictDetectionStore'
+import type { ConflictDetectionResult } from '@/workbench/extensions/manager/types/conflictDetectionTypes'
 
 describe('useConflictDetectionStore', () => {
   beforeEach(() => {
